DEV: port full-width component to be included directly (#126)
Removes the [full-width
component](https://github.com/discourse/discourse-full-width-component)
dependency by adding most of the code directly to the theme. There was
one grid definition in the CSS I ended up removing because it was
conflicting with an existing override in the theme.

Anyone who has previously installed the theme should manually remove the
full-width component from it.